gomod: extract test_utils.py from test_main.py
Move utility function tests into  test_utils.py module
to mirror the utils.py core split.
