+++ tests/unit/package_managers/gomod/test_main.py
@@ -44,10 +44,8 @@ from hermeto.core.package_managers.gomod.main import (
     _get_go_work_path,
     _get_gomod_version,
     _get_repository_name,
-    _go_exec_env,
     _go_list_deps,
     _list_installed_toolchains,
-    _list_toolchain_files,
     _parse_go_sum,
     _parse_local_modules,
     _parse_packages,
@@ -2256,55 +2254,6 @@ def test_get_go_work_path(
     assert mock_go.call_args[0][1] == {"cwd": rooted_tmp_path}
 
 
-@pytest.mark.parametrize(
-    "dir_path, files, expected",
-    [
-        pytest.param(
-            "pkg/mod/cache/download/golang.org/toolchain/@v",
-            ["v0.0.1-go1.21.5.linux-amd64.zip", "v0.0.1-go1.22.0.linux-amd64.zip"],
-            ["v0.0.1-go1.21.5.linux-amd64.zip", "v0.0.1-go1.22.0.linux-amd64.zip"],
-            id="toolchain_files",
-        ),
-        pytest.param(
-            "pkg/mod/cache/download/sumdb/sum.golang.org/lookup",
-            [
-                "golang.org/toolchain@v0.0.1-go1.21.5.linux-amd64",
-                "github.com/example/module@v1.0.0",
-            ],
-            ["golang.org/toolchain@v0.0.1-go1.21.5.linux-amd64"],
-            id="mixed_sumdb_files",
-        ),
-        pytest.param(
-            "pkg/mod/cache/download/golang.org/x/crypto/@v",
-            ["v0.1.0.mod", "v0.1.0.zip"],
-            [],
-            id="golang_org_module_files",
-        ),
-        pytest.param(
-            "pkg/mod/cache/download/github.com/example/module/@v",
-            ["v1.0.0.zip", "v1.1.0.zip"],
-            [],
-            id="other_module_files",
-        ),
-        pytest.param(
-            "foo/bar/golang.org/toolchain/@v",
-            ["version.zip"],
-            [],
-            id="toolchain_under_strange_path",
-        ),
-        pytest.param(
-            "pkg/mod/cache/download",
-            [],
-            [],
-            id="empty_files_list",
-        ),
-    ],
-)
-def test_ignore_toolchain_files(dir_path: str, files: list[str], expected: list[str]) -> None:
-    result = _list_toolchain_files(dir_path, files)
-    assert result == expected
-
-
 class TestGoWork:
     def test_init(
         self,
@@ -2394,38 +2343,3 @@ class TestGoWork:
 
         mock_go.assert_called_once()
         assert mock_go.call_args[0][0] == ["work", "edit", "-json"]
-
-
-_ENV_VARS_BASE_INIT = {v: "/some/path" for v in ("PATH", "HOME", "NETRC")}
-
-
-@pytest.mark.parametrize(
-    "env, extra_env, expected",
-    [
-        pytest.param(_ENV_VARS_BASE_INIT, None, _ENV_VARS_BASE_INIT, id="vars_inherited"),
-        pytest.param(
-            {},
-            None,
-            {"PATH": "", "HOME": "/mocked/home", "NETRC": ""},
-            id="vars_defaults",
-        ),
-        pytest.param(
-            _ENV_VARS_BASE_INIT,
-            {"GOPATH": "/tmp/go"},
-            _ENV_VARS_BASE_INIT | {"GOPATH": "/tmp/go"},
-            id="with_extra_env",
-        ),
-    ],
-)
-@mock.patch("pathlib.Path.home", return_value=Path("/mocked/home"))
-def test_go_exec_env(
-    mock_home: mock.Mock,
-    monkeypatch: pytest.MonkeyPatch,
-    env: dict[str, str],
-    extra_env: dict[str, str] | None,
-    expected: dict[str, str],
-) -> None:
-    monkeypatch.setattr(os, "environ", env)
-
-    actual = _go_exec_env() if extra_env is None else _go_exec_env(**extra_env)
-    assert actual == expected
